handlers/mindmap/requirements.txt
grpcio
protobuf
pydantic
pyyaml
